Move Memo into zcash_primitives::memo
