+++ zcash_client_backend/src/decrypt.rs
@@ -2,7 +2,8 @@ use std::collections::HashMap;
 
 use zcash_primitives::{
     consensus::{self, BlockHeight},
-    note_encryption::{try_sapling_note_decryption, try_sapling_output_recovery, Memo},
+    memo::Memo,
+    note_encryption::{try_sapling_note_decryption, try_sapling_output_recovery},
     primitives::{Note, PaymentAddress},
     transaction::Transaction,
     zip32::ExtendedFullViewingKey,
