+++ zcash_primitives/src/transaction/builder.rs
@@ -14,8 +14,9 @@ use crate::{
     consensus::{self, BlockHeight},
     keys::OutgoingViewingKey,
     legacy::TransparentAddress,
+    memo::Memo,
     merkle_tree::MerklePath,
-    note_encryption::{Memo, SaplingNoteEncryption},
+    note_encryption::SaplingNoteEncryption,
     primitives::{Diversifier, Note, PaymentAddress},
     prover::TxProver,
     redjubjub::PrivateKey,
